% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annamantsoki-package.R
\docType{package}
\name{annamantsoki-package}
\alias{annamantsoki}
\alias{annamantsoki-package}
\title{annamantsoki: Build A Package For Certification Level 2 Exam}
\description{
The package contains functions to calculate body mass index and read all tabs of an excel file.
}
\author{
\strong{Maintainer}: Anna Mantsoki \email{anna.mantsoki@finddx.org}

Other contributors:
\itemize{
  \item FIND [copyright holder]
}

}
\keyword{internal}
