% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{bmi A function to calculate the body mass index}
\usage{
bmi(mass = 80, height = 1.8)
}
\arguments{
\item{mass}{the mass of a person in Kg (type = double, numeric)}

\item{height}{the height of a person in metres (type = double, numeric)}
}
\value{
A vector with the calculated body mass index
}
\description{
bmi A function to calculate the body mass index
}
\examples{
bmi(mass = 80, height = 1.80)
library(dplyr)
data("data_weight")
data_weight \%>\%
  mutate(bmi = bmi(mass = mass, height = height))
}
