% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_weight.R
\docType{data}
\name{data_weight}
\alias{data_weight}
\title{Data for height and mass of Starwars heroes}
\format{
A data frame with 87 rows and 3 variables:
\describe{
\item{name}{name of the hero}
\item{height}{height of the hero, in m}
\item{mass}{mass of the hero, in Kg}
...
}
}
\source{
\url{https://dplyr.tidyverse.org/reference/starwars.html}
}
\usage{
data_weight
}
\description{
A dataset containing information for height and mass of Starwars heroes
}
\keyword{datasets}
