% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_import_excel.R
\name{multi_import_excel}
\alias{multi_import_excel}
\title{bmi A function to read all available tabs from an excel file}
\usage{
multi_import_excel(file = "inst/corrupted_toydatasets.xlsx")
}
\arguments{
\item{file}{the name or pathfile of the excel file (type = character)}
}
\value{
A list of data frames named after the excel tabs
}
\description{
bmi A function to read all available tabs from an excel file
}
\examples{
file <- system.file("corrupted_toydatasets.xlsx", package = "annamantsoki")
multi_import_excel(file)

}
