% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_read_excel.R
\name{safe_read_excel}
\alias{safe_read_excel}
\title{bmi A function to read safely an excel file}
\usage{
safe_read_excel(...)
}
\arguments{
\item{file}{the name or pathfile of the excel file (type = character)}
}
\value{
A data frame or NULL in case of corruption
}
\description{
bmi A function to read safely an excel file
}
